==> Scripts/1. Consistency checks and descriptive analysis.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

path = r'C:\Users\salom\OneDrive\Documentos\11-2025 Instacart Basket Analysis'

path

## Orders dataframe consistency checks

# Import orders data
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# Display the first rows of the dataframe
df_ords.head()

# Display the end rows of the dataframe
df_ords.tail()

# Return the dimensions of the dataframe
df_ords.shape

# Return basic information of the dataframe
df_ords.info()

# Check the data types
df_ords.dtypes

# Check the columns titles
df_ords.columns

# Generate descriptive statistics
df_ords.describe()

# Dropping eval_set column from orders.csv
df_ords.drop(columns = ['eval_set'])

df_ords['days_since_prior_order'].value_counts(dropna = False)

df_ords.describe()

#### In the days_since_prior_order column, the count is smaller than all other columns, which may be due to missing values. The count of order_id matches the max number of orders, indicating that each order is unique. The order_dow goes from 0 to 6, for 7 days, and the order_hour_of_day goes from 0 to 23, for 24 hours, which seems correct. The order number seems to go only from 1 to 100, indicating that the dataframe only stores a maximum of 100 orders per user. 

# Check for mixed-type data in your df_ords dataframe.

for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].map(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

#### No column was found with mixed-type data

# Run a check for missing values in your df_ords dataframe.

df_ords.isnull().sum()

# View missing values in a subset

df_nan2 = df_ords[df_ords['days_since_prior_order'].isnull() == True]

df_nan2

#### As indicated by the descriptive statistics, there are 206209 rows with missing values in the days_since_prior_order column. These all appear to correspond to the first order placed by each user, therefore, there was no prior order to compare to.

df_ords.describe()

####  Missing values were left as is, since they all appear to correspond to the first order placed by each user.

# Run a check for duplicate values in your df_ords_clean data

df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups.shape

#### There are no duplicates, as indicated by the descriptive statistics showing 3421083 unique order numbers

# Export clean dataframe

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_checked.csv'))

## Products dataframe consistency checks

# Import products data
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# Print the first 20 rows of the df_prods dataframe.
df_prods.head(20)

# Print the the last 35 rows of the df_prods dataframe.
df_prods.tail(35)

# Print the names of the columns in the df_prods dataframe
df_prods.columns

# Print the number of rows and columns in the df_prods dataframe
df_prods.shape

# Answer the following questions using the relevant commands
# What’s the maximum value of the “aisle_id” column?
df_prods.describe()

# The maximum value of the “aisle_id” column is 134

# What’s the data type of the “department_id” column?
df_prods.dtypes

# The data type of the “department_id” column is integer

# Count missing values

df_prods.isnull().sum()

# View missing values in a subset

df_nan = df_prods[df_prods['product_name'].isnull() == True]

# Check number of rows in current products dataframe

df_prods.shape

df_prods.describe()

# Create new clean dataframe without missing values

df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# Check number of rows in new products dataframe

df_prods_clean.shape

df_prods.dropna(inplace = False)

# New subset with full duplicates

df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

# Drop duplicates in new dataframe

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups

df_prods_clean_no_dups.describe()

df_prods_clean_no_dups.sort_values(by=['prices']).tail(5)

# Impute inaccurate values with median price (7.1)

df_prods_clean_no_dups.loc[df_prods_clean_no_dups['product_id']==21553, 'prices'] = 7.1

df_prods_clean_no_dups.loc[df_prods_clean_no_dups['product_id']==33664, 'prices'] = 7.1

df_prods_clean_no_dups[df_prods_clean_no_dups['product_id']==33664]

df_prods_clean_no_dups[df_prods_clean_no_dups['product_id']==21553]

df_prods_clean_no_dups.describe()

# Export clean dataframe

df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'))
