==> Scripts/2. Wrangling procedures.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

path = r'C:\Users\salom\OneDrive\Documentos\11-2025 Instacart Basket Analysis'

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

## Practice wrangling procedures

# Dropping eval_set column from orders.csv
df_ords.drop(columns = ['eval_set'])

df_ords = df_ords.drop(columns = ['eval_set'])

df_ords['days_since_prior_order'].value_counts(dropna = False)

# Rename column order_dow
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# check 
df_ords.head()

# Changing order_id data type to string
df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['order_id'].dtype

# Importing departments dataset
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

df_dep.head()

# Transpose departments dataset
df_dep.T

df_dep_t = df_dep.T

df_dep_t

# Adding index 
df_dep_t.reset_index()

# Create new header
new_header = df_dep_t.iloc[0]

new_header 

# Copy first row onward
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

df_dep_t_new.columns = new_header #Set the header rows as the df header

df_dep_t_new

# Turn df_dep into a data dictionary 
data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

print(data_dict.get('19'))

# Create snacks department subset
df_snacks =  df_prods[df_prods['department_id']==19]

# Search snacks dept data
df_prods['department_id']==19

# List true values within df_prods 
df_prods[df_prods['department_id']==19]

df_snacks.head()

# look for instances where the value of "department_id" is 19 within the df_prods database
df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

df_snacks_2.head()

# Look into a list (several values)
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([17,18,19])]

df_snacks_3.head()

## Exercise 4.4 Task. Additional wrangling procedures.

# Find another numeric variable in the df_ords dataframe that doesn’t need to be included in your analysis and change it to a suitable format.
df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords['user_id'].dtype

# Look for a variable in your df_ords dataframe with an unintuitive name and change its name without overwriting the dataframe.
df_ords.rename(columns = {'orders_day_of_week' : 'day_of_order', 'order_hour_of_day' : 'time_of_order'})

df_ords.head()

# Your client wants to know what the busiest hour is for placing orders. Find the frequency of the corresponding variable and share your findings.
df_ords['order_hour_of_day'].value_counts(dropna = False)

# The busiest hour of day for placing orders is 10 am.

# Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary.
print(data_dict.get('4'))

# The sales team in your client’s organization wants to know more about breakfast item sales. Create a subset containing only the required information.
df_breakfast =  df_prods[df_prods['department_id']==14]

df_breakfast.head()

# They’d also like to see details about products that customers might use to throw dinner parties. 
# Your task is to find all observations from the entire dataframe that include items from the following departments: alcohol, deli, beverages, and meat/seafood. 
# You’ll need to present this subset to your client.
df_dinner_parties = df_prods.loc[df_prods['department_id'].isin([5,20,7,12])]

df_dinner_parties.head(30)

# It’s important that you keep track of total counts in your dataframes. How many rows does the last dataframe you created have?
df_dinner_parties.shape

# The dinner parties subset has 7650 rows.

# Someone from the data engineers team in Instacart thinks they’ve spotted something strange about the customer with a "user_id" of “1.” - I'm unsure if the "." was a mistake. There were no values for "1.", so I did the analysis with value "1".
# Extract all the information you can about this user.
df_ords['user_id']=='1'

df_ords[df_ords['user_id']=='1']

df_user_1 = df_ords[df_ords['user_id']=='1']

#Generate descriptive statistics of user '1' orders.
df_user_1.describe()

## Export dataframes

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
