==> Scripts/3. Combining & Exporting Data.ipynb <==
# Import libraries

import pandas as pd
import numpy as np
import os

### Import dataframes

path = r'C:\Users\salom\OneDrive\Documentos\11-2025 Instacart Basket Analysis'

# Import new dataframe

df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)

# Check the dimensions of the imported dataframe

df_ords_prior.shape

# Check the output

df_ords_prior.head()

# Count missing values

df_ords_prior.isnull().sum()

# New subset with full duplicates

df_dups = df_ords_prior[df_ords_prior.duplicated()]

df_dups

#Import orders dataframe

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'), index_col = False)

# Check the dimensions

df_ords.shape

# Check the output

df_ords.head()

## Merge dataframes

df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

# Check the output

df_merged_large.head()

# Sum up all the values in the “_merge” column

df_merged_large['_merge'].value_counts()

# Remove unwanted columns

df_merged_large = df_merged_large.drop(columns = ['Unnamed: 0', '_merge'])

df_merged_large.shape

### Export merged dataframe

# Export data to pkl

df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))

# Import checked products dataframe

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col = False)

# Check output

df_prods.head()

df_prods.describe()

# Remove unwanted columns

df_prods = df_prods.drop(columns = ['Unnamed: 0'])

###  Combine the orders_products_combined dataframe with your products data set

# Confirm the results of the merge using the merge flag

pd.merge(df_merged_large,df_prods, on = ['product_id'], indicator = True)

df_ords_prods_merged = df_merged_large.merge(df_prods, on = 'product_id')

df_ords_prods_merged.head()

df_ords_prods_merged.shape

# Export data to pkl

df_ords_prods_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merge.pkl'))
