==> Scripts/4. Deriving New Variables.ipynb <==
# Import libraries

import pandas as pd
import numpy as np
import os

path = r'C:\Users\salom\OneDrive\Documentos\11-2025 Instacart Basket Analysis'

# Import merged pkl dataframe

df_ords_prods_merged = pd.read_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merge.pkl'))

# Created subset of the dataframe for the first one million rows.

df = df_ords_prods_merged[:1000000]

df.shape

df.head()

df.tail()

# Define price label function

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Apply the function

df['price_range'] = df.apply(price_label, axis=1)

df.value_counts('price_range')

df['prices'].max()

# Create conditions using loc function

df_ords_prods_merged.loc[df_ords_prods_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods_merged.loc[(df_ords_prods_merged['prices'] <= 15) & (df_ords_prods_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df_ords_prods_merged.loc[df_ords_prods_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods_merged['price_range_loc'].value_counts(dropna=False)

# Print the frequency of the “orders_day_of_week” column

df_ords_prods_merged['orders_day_of_week'].value_counts(dropna=False)

# Create list for busiest day using a for loop

result = []

for value in df_ords_prods_merged['orders_day_of_week']:
  if value == 0:
    result.append('Busiest day')
  elif value == 4:
    result.append('Least busy')
  else:
    result.append('Regularly busy')

result

# Create new column to combine with list for busiest day

df_ords_prods_merged['busiest_day']= result

# Print the frequency of this new column and cross-check it with the frequency of “orders_day_of_week”

df_ords_prods_merged['busiest_day'].value_counts(dropna=False)

### Exercise 4.7 Directions

#### Suppose your clients have changed their minds about the labels you created in your “busiest_day” column. Now, they want “Busiest day” to become “Busiest days” (plural). This label should correspond with the two busiest days of the week as opposed to the single busiest day. At the same time, they’d also like to know the two slowest days. Create a new column for this using a suitable method.

# Create new list for busiest days using a for loop

result = []

for value in df_ords_prods_merged['orders_day_of_week']:
  if value == 0:
    result.append('Busiest days')
  elif value == 1:
    result.append('Busiest days')
  elif value == 4:
    result.append('Slowest days')
  elif value == 3:
    result.append('Slowest days')
  else:
    result.append('Regularly busy')

result

# Create new column to combine with list for busiest days

df_ords_prods_merged['busiest_days']= result

# Print the frequency of this new column and cross-check it with the frequency of “orders_day_of_week”

df_ords_prods_merged['busiest_days'].value_counts(dropna=False)

#### In this new column the number of regularly busy days is reduced by more than half, while the busiest and slowest days double. The values sum is accurate. 

#### When too many users make Instacart orders at the same time, the app freezes. The senior technical officer at Instacart wants you to identify the busiest hours of the day. Rather than by hour, they want periods of time labeled “Most orders,” “Average orders,” and “Fewest orders.” Create a new column containing these labels called “busiest_period_of_day.”

# Create a temp table to determine the frequency of the “order_hour_of_day” column

orders_frequency = df_ords_prods_merged['order_hour_of_day'].value_counts(dropna=False)

orders_frequency

orders_frequency.describe()

# Create conditions using loc function: 
# Most orders -q4- 10,11,14,15,13,12
# Average orders -q3,q2- 16,9,17,8,18,19,20,7,21,22,23,6
# Fewest orders -q1- 0,1,5,2,4,3

df_ords_prods_merged.loc[(df_ords_prods_merged['order_hour_of_day'] >= 10) & (df_ords_prods_merged['order_hour_of_day'] <= 15), 'busiest_period_of_day'] = 'Most orders'

df_ords_prods_merged.loc[(df_ords_prods_merged['order_hour_of_day'] >= 6) & (df_ords_prods_merged['order_hour_of_day'] <= 9), 'busiest_period_of_day'] = 'Average orders'

df_ords_prods_merged.loc[(df_ords_prods_merged['order_hour_of_day'] >= 16) & (df_ords_prods_merged['order_hour_of_day'] <= 23), 'busiest_period_of_day'] = 'Average orders'

df_ords_prods_merged.loc[(df_ords_prods_merged['order_hour_of_day'] <= 5), 'busiest_period_of_day'] = 'Fewest orders'

# Print the frequency of this new column

df_ords_prods_merged['busiest_period_of_day'].value_counts(dropna=False)

df_ords_prods_merged.head()

# Export data to pkl

df_ords_prods_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_derived.pkl'))
