==> Scripts/5. Grouping Data & Aggregating Variables.ipynb <==
# Import libraries

import pandas as pd
import numpy as np
import os

path = r'C:\Users\salom\OneDrive\Documentos\11-2025 Instacart Basket Analysis'

# Import last pkl dataframe

ords_prods_merged = pd.read_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_derived.pkl'))

df = ords_prods_merged[:1000000]

df.shape

df.groupby('product_name')

# Aggregating Data with agg()

df.groupby('department_id').agg({'order_number': ['mean']})

# Conduct aggregations without use of the agg() function

df.groupby('department_id')['order_number'].mean()

df.groupby('department_id').order_number.mean()

# Performing Multiple Aggregations

df.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

# Create new “max_order” column

ords_prods_merged['max_order'] = ords_prods_merged.groupby(['user_id'])['order_number'].transform(np.max)

ords_prods_merged.head(500)

# Deriving Columns with loc()

ords_prods_merged.loc[ords_prods_merged['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

ords_prods_merged.loc[(ords_prods_merged['max_order'] <= 40) & (ords_prods_merged['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

ords_prods_merged.loc[ords_prods_merged['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# Print the frequency of new “loyalty_flag” column

ords_prods_merged.value_counts('loyalty_flag')

# Check only relevant columns

ords_prods_merged[['user_id', 'loyalty_flag', 'order_number']].tail(500)

### Find the aggregated mean of the “order_number” column grouped by “department_id” for the entire dataframe.

ords_prods_merged.groupby('department_id').agg({'order_number': ['mean']})

### Analyze the result. How do the results for the entire dataframe differ from those of the subset? 

#### The average number of orders by department increases slightly (at least by 1) when aggregating the entire dataset.

### The marketing team at Instacart wants to know whether there’s a difference between the spending habits of the three types of customers you identified. Use the loyalty flag you created and check the basic statistics of the product prices for each loyalty category (Loyal Customer, Regular Customer, and New Customer). What you’re trying to determine is whether the prices of products purchased by loyal customers differ from those purchased by regular or new customers.

ords_prods_merged.groupby('loyalty_flag').agg({'prices': ['mean', 'min', 'max']})

#### Average product prices are similar for all loyalty categories.

### The team now wants to target different types of spenders in their marketing campaigns. This can be achieved by looking at the prices of the items people are buying. Create a spending flag for each user based on the average price across all their orders using the following criteria: 
#### If the mean of the prices of products purchased by a user is lower than 10, then flag them as a “Low spender.” 
#### If the mean of the prices of products purchased by a user is higher than or equal to 10, then flag them as a “High spender.”

# Create new “avg_price_by_user” column

ords_prods_merged['avg_price_by_user'] = ords_prods_merged.groupby(['user_id'])['prices'].transform(np.mean)

# Check only relevant columns

ords_prods_merged[['user_id', 'avg_price_by_user']].describe()

# Create spending flag with loc()

ords_prods_merged.loc[ords_prods_merged['avg_price_by_user'] >= 10, 'spending_flag'] = 'High spender'

ords_prods_merged.loc[ords_prods_merged['avg_price_by_user'] < 10, 'spending_flag'] = 'Low spender'

# Print the frequency of new “spending_flag” column

ords_prods_merged.value_counts('spending_flag')

# Check only relevant columns

ords_prods_merged[['user_id', 'avg_price_by_user', 'spending_flag']].head(1000)

### In order to send relevant notifications to users within the app (for instance, asking users if they want to buy the same item again), the Instacart team wants you to determine frequent versus non-frequent customers. Create an order frequency flag that marks the regularity of a user’s ordering behavior according to the median in the “days_since_prior_order” column. The criteria for the flag should be as follows:
#### If the median of “days_since_prior_order” is higher than 20, then the customer should be labeled a “Non-frequent customer.”
#### If the median is higher than 10 and lower than or equal to 20, then the customer should be labeled a “Regular customer.”
#### If the median is lower than or equal to 10, then the customer should be labeled a “Frequent customer.”

# Create new “avg_days_prior_order” column

ords_prods_merged['avg_days_prior_order'] = ords_prods_merged.groupby(['user_id'])['days_since_prior_order'].transform(np.mean)

# Check only relevant columns

ords_prods_merged[['user_id', 'avg_days_prior_order']].describe()

# Create frequency flag with loc()

ords_prods_merged.loc[ords_prods_merged['avg_days_prior_order'] > 20, 'frequency_flag'] = 'Non-frequent customer'

ords_prods_merged.loc[(ords_prods_merged['avg_days_prior_order'] <= 20) & (ords_prods_merged['avg_days_prior_order'] > 10), 'frequency_flag'] = 'Regular customer'

ords_prods_merged.loc[ords_prods_merged['avg_days_prior_order'] <= 10, 'frequency_flag'] = 'Frequent customer'

# Print the frequency of the “frequency_flag” column

ords_prods_merged.value_counts('frequency_flag')

# Check only relevant columns

ords_prods_merged[['user_id', 'avg_days_prior_order', 'frequency_flag']].head(1000)

ords_prods_merged.head(5)

# Export data to pkl

ords_prods_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_aggregated.pkl'))
