==> Scripts/6. Customer data checks & merge.ipynb <==
# Import libraries

import pandas as pd
import numpy as np
import os

path = r'C:\Users\salom\OneDrive\Documentos\11-2025 Instacart Basket Analysis'

df_custs = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'), index_col = False)

df_custs.shape

df_custs.head(10)

df_custs.dtypes

df_custs.describe()

#### The count of user_id matches the count of values in other columns, indicating that each row is unique. The age range is between 18 and 81 years, with a mean of 49, which seems within a reasonable range. The number of dependants and the income range seem reasonable values as well.

# Count values for object columns
df_custs.value_counts('STATE')

#### Customers are evenly distributed across states.

df_custs.value_counts('Gender')

#### There are slightly more male than female customers.

df_custs.value_counts('fam_status')

#### There are significantly more married customers than any other family status.

df_custs.value_counts('date_joined')

# Count missing values

df_custs.isnull().sum()

# View missing values in a subset

df_nan = df_custs[df_custs['First Name'].isnull() == True]

df_nan

#### Because the missing first name does not affect our analysis, these rows should not be removed. An 'Unknown' value can be used to impute the null values.

# Impute missing values with 'Unknown'

df_custs['First Name'] = df_custs['First Name'].fillna('Unknown')

df_custs.head(50)

df_custs.isnull().sum()

# Check for mixed-type data

for col in df_custs.columns.tolist():
  weird = (df_custs[[col]].map(type) != df_custs[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_custs[weird]) > 0:
    print (col)

#### No column was found with mixed-type data

# Run a check for duplicate values

df_custs_dups = df_custs[df_custs.duplicated()]

df_custs_dups.shape

#### There are no duplicates, as indicated by the descriptive statistics showing 206209 unique order numbers

# Rename columns

df_custs.rename(columns = {'First Name' : 'first_name', 'Surnam' : 'last_name', 'Gender' : 'gender', 'STATE' : 'state', 'Age' : 'age'}, inplace = True)

df_custs.head(5)

# Changing user_id, first_name, last_name, gender, state, fam_status data type to string

df_custs['user_id'] = df_custs['user_id'].astype('str')
df_custs['first_name'] = df_custs['first_name'].astype('str')
df_custs['last_name'] = df_custs['last_name'].astype('str')
df_custs['gender'] = df_custs['gender'].astype('str')
df_custs['state'] = df_custs['state'].astype('str')
df_custs['fam_status'] = df_custs['fam_status'].astype('str')

df_custs.dtypes

df_custs['user_id'].dtype

#Export clean dataframe

df_custs.to_csv(os.path.join(path, '02 Data','Prepared Data', 'customers_clean.csv'))

## Combine your customer data with the rest of your prepared Instacart data.

# Import last pkl dataframe

ords_prods = pd.read_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_aggregated.pkl'))

ords_prods.value_counts('user_id')

#### Count of unique customers coincides for both dataframes.

# Changing user_id data type in aggregated dataframe to string

ords_prods['user_id'] = ords_prods['user_id'].astype('str')

ords_prods['user_id'].dtype

# Drop unnecessary columns for the merged dataframe. 
# Since every user has a unique ID, first and last names are not necessary for the analysis. 
# Date joined is also not necessary for the analysis of combined dataframes.

df_custs_m = df_custs.drop(columns = ['first_name', 'last_name', 'date_joined'])


df_custs_m.head()

# Merge dataframes 

ords_prods_custs = ords_prods.merge(df_custs_m, on = 'user_id', indicator = True)

ords_prods_custs.head(10)

# Remove unwanted columns

ords_prods_custs = ords_prods_custs.drop(columns = ['_merge'])

ords_prods_custs.shape

ords_prods_custs.describe()

# Export data to pkl

ords_prods_custs.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_customers.pkl'))
